Source Code/server/requirements.txt
firebase-admin>=6.2.0
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=4.9.0
pandas>=2.0.0
flask>=3.0.0
flask-cors>=4.0.0
python-dateutil>=2.8.2